feat: sqs client refactoring
