+++ crates/worker/src/sqs_listener.rs
@@ -8,16 +8,17 @@ use std::time::Duration;
 use tokio::task::JoinHandle;
 use tokio::time::sleep;
 
+use crate::sqs_client::wrapper::SqsClientWrapper;
 use crate::sqs_client::SqsClient;
 
 pub struct SqsListener {
     handle: JoinHandle<Result<(), SqsReceiveError>>,
     receiver: Receiver<Message>,
-    client: SqsClient,
+    client: SqsClientWrapper,
 }
 
 impl SqsListener {
-    pub fn new(client: SqsClient, poll_interval: Duration) -> Self {
+    pub fn new(client: SqsClientWrapper, poll_interval: Duration) -> Self {
         // TODO: unbounded?
         let (sender, receiver) = async_channel::bounded(1000);
         let handle = tokio::spawn(Self::listen(client.clone(), sender, poll_interval));
@@ -30,7 +31,7 @@ impl SqsListener {
     }
 
     async fn listen(
-        client: SqsClient,
+        client: SqsClientWrapper,
         sender: Sender<Message>,
         poll_interval: Duration,
     ) -> Result<(), SdkError<ReceiveMessageError, HttpResponse>> {
@@ -65,7 +66,7 @@ impl SqsListener {
 }
 
 pub struct SqsReceiver {
-    client: SqsClient,
+    client: SqsClientWrapper,
     receiver: Receiver<Message>,
 }
